Fix Windows Dropdown Background
The dropdown menus on Windows have looked messed up for a while. The background coloring has always been white despite changing themes, while the text changes color with the theme. Now the Background of the dropdown menus mirror the background of the app. I have not tested on Mac or Linux.